+++ src/Ractor/Connections.fs
@@ -12,18 +12,17 @@ open Ractor
 // identical everywhere but the actor behavior should depend on the tags in the environment
 // TODO test high CPU load and many (1000+) IO-like actors
 
-type Ractor(redisConnectionString : string) = 
-    let redis = Redis(redisConnectionString, "R")
+type Connections(redisConnectionString : string) = 
+    //let redis = Redis(redisConnectionString, "R")
        
     static let dbs = Dictionary<string, IPocoPersistor>()
     static let blobs = Dictionary<string, IBlobPersistor>()
     static let redises = Dictionary<string, Redis>()  
 
-    do
-        redis.Serializer <- JsonSerializer()
-
-        if String.IsNullOrEmpty(ActorImpl<_,_>.DefaultRedisConnectionString) then
-            ActorImpl<_,_>.DefaultRedisConnectionString <- redisConnectionString
+//    do
+//        redis.Serializer <- JsonSerializer()
+//        if String.IsNullOrEmpty(ActorImpl<_,_>.DefaultRedisConnectionString) then
+//            ActorImpl<_,_>.DefaultRedisConnectionString <- redisConnectionString
 
     static let mutable logger = 
 #if DEBUG
@@ -38,7 +37,7 @@ type Ractor(redisConnectionString : string) =
             ActorBase.Logger <- newLogger
             logger <- newLogger
 
-    static member RegisterDB(persistor : IPocoPersistor) = Ractor.RegisterDB(persistor, "")
+    static member RegisterDB(persistor : IPocoPersistor) = Connections.RegisterDB(persistor, "")
     
     static member RegisterDB(persistor : IPocoPersistor, id : string) = 
         let id = id.ToLowerInvariant()
@@ -52,13 +51,13 @@ type Ractor(redisConnectionString : string) =
     static member GetDB() = dbs.[""]
     
     static member RegisterBlobStorage(persistor : IBlobPersistor, id : string) =
-        let id = id.ToLowerInvariant() 
+        let id = id.ToLowerInvariant()
         if blobs.ContainsKey(id) then 
             raise (InvalidOperationException("Blob Storage with the same id already exists: " + id))
         blobs.Add(id, persistor)
         persistor
     
-    static member RegisterBlobStorage(persistor : IBlobPersistor) = Ractor.RegisterBlobStorage(persistor, "")
+    static member RegisterBlobStorage(persistor : IBlobPersistor) = Connections.RegisterBlobStorage(persistor, "")
     static member GetBlobStorage(id : string) = 
         let id = id.ToLowerInvariant() 
         blobs.[id]
@@ -70,7 +69,7 @@ type Ractor(redisConnectionString : string) =
         redises.Add(id, redis)
         redis
     
-    static member RegisterRedis(redis : Redis) = Ractor.RegisterRedis(redis, "")
+    static member RegisterRedis(redis : Redis) = Connections.RegisterRedis(redis, "")
     static member GetRedis(id : string) = 
         let id = id.ToLowerInvariant() 
         redises.[id]