+++ tests/Ractor.Tests/Basics.fs
@@ -16,7 +16,7 @@ open FsUnit
 
 type Greeter ()=
     inherit Actor<string,string>()
-    override this.Redis = "127.0.0.1:6379,resolveDns=true"
+    override this.RedisConnectionString = "127.0.0.1:6379,resolveDns=true"
     override this.Computation(input) : Async<string> =  
         async {
                 Console.WriteLine("Hello, " + input)
@@ -27,7 +27,7 @@ type Greeter ()=
 
 type Echo () =
     inherit Actor<string,string>()
-    override this.Redis = "127.0.0.1:6379,resolveDns=true"
+    override this.RedisConnectionString = "127.0.0.1:6379,resolveDns=true"
     override this.Computation(input) : Async<string> = 
         async { 
             Console.WriteLine("Echo: " + input)
@@ -37,7 +37,7 @@ type Echo () =
 
 type Incrementer ()=
     inherit Actor<int,int>()
-    override this.Redis = "127.0.0.1:6379,resolveDns=true"
+    override this.RedisConnectionString = "127.0.0.1:6379,resolveDns=true"
     override this.Computation(input) : Async<int> =  
         async {
                 //Console.WriteLine("Incremented to: " + (input + 1).ToString())
@@ -51,8 +51,6 @@ type Incrementer ()=
 [<Test>]
 let ``Hello, Actors`` () =
     
-    let fredis = new Ractor("127.0.0.1:6379,resolveDns=true") // TODO this is ugly, make module Ractor
-
     let greeter = Greeter()
 
     // we could use new instance each time for shorter syntax
@@ -117,9 +115,7 @@ let ``Iterate`` () =
 
 [<Test>]
 let ``PostAndReply local execution`` () =
-    let fredis = new Ractor("localhost")
     
-
     let greeter = Greeter()
     
     greeter.Start()